requirements.txt
alfred==0.3
bbox==0.9.2
coremltools==5.1.0
Cython==0.29.24
Flask==2.0.1
ipython==7.30.1
keras==2.7.0
keras_nightly==2.5.0.dev2021032900
matplotlib==3.5.0
nb==0.1.2
numpy==1.21.2
opencv-python-headless
pandas==1.3.4
Pillow==8.4.0
PyYAML==6.0
requests==2.25.1
scikit_image==0.18.3
scipy==1.7.1
seaborn==0.11.2
thop==0.0.31.post2005241907
tqdm==4.61.1
Werkzeug==2.0.2